+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,6 @@
         </activity>
         <activity android:name=".tutorial.TutorialStartActivity"/>
         <activity android:name=".tutorial.Tutorial1StepActivity" android:windowSoftInputMode="adjustNothing" />
-        <activity android:name=".DatePickerActivity" android:theme="@android:style/Theme.Dialog"/>
         <activity android:name=".tutorial.Tutorial2StepActivity" android:windowSoftInputMode="adjustNothing" />
         <activity android:name=".MypageActivity"/>
         <activity android:name=".self.SelfMainActivity"/>
