+++ app/src/main/java/kr/co/healthcare/tutorial/Tutorial1StepActivity.java
@@ -4,12 +4,13 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
+
 import android.widget.EditText;
 import android.widget.RadioGroup;
 
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.DialogFragment;
 
-import kr.co.healthcare.DatePickerActivity;
 import kr.co.healthcare.exception.NoneBirthSelectedException;
 import kr.co.healthcare.exception.NoneInputException;
 import kr.co.healthcare.exception.NoneSexSelectedException;
@@ -25,49 +26,38 @@ public class Tutorial1StepActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_tutorial_1_step);
-
-        //생년월일 버튼 클릭 시
-        dateBtn = (Button) findViewById(R.id.dateBtn);
-        dateBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(Tutorial1StepActivity.this, DatePickerActivity.class);
-                startActivityForResult(intent, ACT_SET_BIRTH);
-            }
-        });
-
-        //첫 번째 튜토리얼 완료 버튼 클릭 시
-        final Button firstStepFinishBtn = (Button) findViewById(R.id.firstStepFinishBtn);
-        firstStepFinishBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                setDataAndChangeView();
-            }
-        });
     }
 
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        if (requestCode == 1) {
-            if (resultCode == RESULT_OK) {
-                //생년월일 데이터 받기
-                String result = data.getIntExtra("mYear", 0) + "-"
-                        + data.getIntExtra("mMonth", 1) + "-"
-                        + data.getIntExtra("mDay", 1);
-                dateBtn.setText(result);
-                isBirthChecked = true;
-            }
-        }
+    //생년월일 버튼 클릭 시
+    public void showDatePicker(View view) {
+        DialogFragment newFragment = new DatePickerFragment();
+        newFragment.show(getSupportFragmentManager(),"datePicker");
+    }
+
+    //첫 번째 튜토리얼 완료 버튼 클릭 시
+    public void firstStepFinish(View view){
+        setDataAndChangeView();
+    }
+
+    public void processDatePickerResult(int year, int month, int day){
+        String year_string = Integer.toString(year);
+        String month_string = Integer.toString(month+1);
+        String day_string = Integer.toString(day);
+
+        PreferenceManger.setString(this,"year", year_string);
+        PreferenceManger.setString(this,"month", month_string);
+        PreferenceManger.setString(this,"day", day_string);
+        isBirthChecked = true;
+
+        final Button dateBtn = (Button) findViewById(R.id.dateBtn);
+        dateBtn.setText(year_string+ "-" + month_string + "-" + day_string);
     }
 
     //이름, 생년월일, 성별 데이터 저장 후 다음 튜토리얼 스텝으로 화면 변경
-
-    void setDataAndChangeView(){
+    private void setDataAndChangeView(){
         EditText etName = (EditText)findViewById(R.id.editTextPersonName);
         RadioGroup rg = (RadioGroup) findViewById(R.id.rdGroup);
         try {
-
             //이름 예외처리: 최소 1글자 이상이여야 함.
             if(etName.getText().toString().length() != 0)
                 PreferenceManger.setString(this,"name", etName.getText().toString());
@@ -98,7 +88,7 @@ public class Tutorial1StepActivity extends AppCompatActivity {
         }
     }
 
-    void changeView(){
+    private void changeView(){
         Intent intent = new Intent(getApplicationContext(), Tutorial2StepActivity.class);
         startActivity(intent);
         overridePendingTransition(0, 0);
