--- app/src/main/res/layout/activity_date_picker.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <DatePicker
-        android:id="@+id/vDatePicker"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:calendarViewShown="false"
-        android:datePickerMode="spinner" />
-
-    <Button
-        android:id="@+id/vDateEnter"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:onClick="mOnClick"
-        android:text="확인" />
-</LinearLayout>
