+++ app/src/main/res/layout/activity_tutorial_1_step.xml
@@ -90,10 +90,10 @@
             android:layout_width="250dp"
             android:layout_height="wrap_content"
             android:layout_marginBottom="20dp"
-            android:singleLine="true"
             android:background="@drawable/edittext_normal"
             android:padding="5dp"
-            android:text="생년월일 선택하기"
+            android:singleLine="true"
+            android:text="생년월일을 선택하세요"
             android:textColor="#292929"
             android:textSize="22sp" />
